Vnmrbg is now 64-bit
Made patch for external testing (Thanks Brian Antalek).
Fixed compiler warnings.

+++ src/psg/oopc.h
@@ -17,7 +17,7 @@
 
 typedef char c68char;
 typedef short c68int;
-typedef long c68long;
+typedef int c68long;
 
 #endif
